pyproject.toml
[project]
name = "rasa-fraym"
version = "0.1.0"
description = "FastAPI application for Rasa Fraym"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "openai",
    "prisma",
    "python-dotenv",
    "python-multipart",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "aiosqlite"
]
